Member Annotation Colors
Successfully got user colors to show based on what they previously selected.
