+++ MapShare/MapShare/Extensions/String+Ext.swift
@@ -5,7 +5,7 @@
 //  Created by iMac Pro on 5/4/23.
 //
 
-import Foundation
+import UIKit
 
 extension String {
     static func generateRandomCode() -> String {
@@ -19,5 +19,16 @@ extension String {
         }
         
         return randomString
+    }
+    
+    static func convertToColorFromString(string: String) -> UIColor? {
+        let components = string.components(separatedBy: ",")
+        guard components.count == 4,
+              let red = Float(components[0]),
+              let green = Float(components[1]),
+              let blue = Float(components[2]),
+              let alpha = Float(components[3]) else { return nil}
+        
+        return UIColor(red: CGFloat(red), green: CGFloat(green), blue: CGFloat(blue), alpha: CGFloat(alpha))
     }
 }
