+++ MapShare/MapShare/Supporting Files/Info.plist
@@ -2,6 +2,8 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
+	<string>We need access to your location when using this app.</string>
 	<key>CFBundleIdentifier</key>
 	<string></string>
 	<key>NSLocationWhenInUseUsageDescription</key>
